frontend: Move author counts into a standalone component.
